Revert "Move CSI code from IndexManager to FieldIndexManagementApiImpl"
This reverts commit 08afa67c76c0b5978f154957391604b725e03930.
